support CGO_ENABLED=0
Add minimal fallbacks when building without CGO.
